move mapResponseForRender to own file
